problem with compiling due to intense use of templates
