+++ main.cpp
@@ -1,4 +1,4 @@
-#include "parser.h"
+#include "launcher_functions.h"
 #include <cassert>
 #include <string>
 #include <vector>
@@ -17,128 +17,6 @@ void test() {
     test_node();
 }
 #endif
-template <env_change_type env_change_type>
-void run_simulation_given_mut_type(const cxxopts::ParseResult& results)
-{
-    auto mut_type = convert_ind_args(results).m_mutation_type;
-
-    if(mut_type == mutation_type::weights)
-    {
-        using net_t = network<mutation_type::weights>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-        observer<sim_t> o;
-
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::activation) {
-
-        using net_t = network<mutation_type::activation>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::weights_and_activation) {
-
-        using net_t = network<mutation_type::weights_and_activation>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::duplication) {
-
-        using net_t = network<mutation_type::duplication>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::NRduplication) {
-
-        using net_t = network<mutation_type::NRduplication>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::addition) {
-
-        using net_t = network<mutation_type::addition>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else if (mut_type == mutation_type::NRaddition) {
-
-        using net_t = network<mutation_type::NRaddition>;
-        using ind_t = individual<net_t>;
-        using pop_t = population<ind_t>;
-        using sim_t = simulation<pop_t, env_change_type>;
-
-        observer<sim_t> o;
-        auto s = create_simulation<pop_t, env_change_type>(results);
-        exec<sim_t>(s, o) ;
-        save_json(o,
-                  create_save_name_from_observer_data(o));
-    }
-    else
-    {
-        throw std::runtime_error{"unknown mutation type"};
-    }
-}
-
-void run_simulation_given_arguments(const cxxopts::ParseResult& results)
-{
-    auto env_change_type = convert_sim_args(results).change_type ;
-    if(env_change_type == env_change_type::asymmetrical)
-    {
-        run_simulation_given_mut_type< env_change_type::asymmetrical>(results);
-    }
-    else if(env_change_type == env_change_type::symmetrical)
-    {
-        run_simulation_given_mut_type<env_change_type::asymmetrical>(results);
-    }
-    else if(env_change_type == env_change_type::regular)
-    {
-        run_simulation_given_mut_type<env_change_type::regular>(results);
-    }
-    else{
-        throw std::runtime_error{"unknown change type"};
-    }
-}
-
 
 int main(int argc, char ** argv) //!OCLINT tests may be long
 {
@@ -158,7 +36,7 @@ int main(int argc, char ** argv) //!OCLINT tests may be long
         assert(1 == 2);
 #endif
 
-        run_simulation_given_arguments(results);
+     //   run_simulation_given_sel_type(results);
 
     }  catch (int exc) {
         if(exc==1)
